# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: API DGAP
  version: 1.0.0
  description: Direction Générale de l'Administration Pénitentiaire — Sénégal. API
    REST v1.
paths:
...
  /api/v1/backoffice/articles:
    get:
      operationId: backoffice_articles_list
      description: CRUD + workflow éditorial des articles — `/api/v1/backoffice/articles`.
      parameters:
      - name: cursor
        required: false
        in: query
        description: La valeur du curseur de pagination.
        schema:
          type: string
      - name: limit
        required: false
        in: query
        description: Nombre de résultats à retourner par page.
        schema:
          type: integer
      tags:
      - backoffice
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedArticleBackofficeList'
          description: ''
    post:
      operationId: backoffice_articles_create
      description: CRUD + workflow éditorial des articles — `/api/v1/backoffice/articles`.
      tags:
      - backoffice
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ArticleBackoffice'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/ArticleBackoffice'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/ArticleBackoffice'
        required: true
      security:
      - jwtAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArticleBackoffice'
          description: ''
...
  /api/v1/backoffice/pages:
    get:
      operationId: backoffice_pages_list
      description: CRUD + workflow éditorial des pages — `/api/v1/backoffice/pages`.
      parameters:
      - name: cursor
        required: false
        in: query
        description: La valeur du curseur de pagination.
        schema:
          type: string
      - name: limit
        required: false
        in: query
        description: Nombre de résultats à retourner par page.
        schema:
          type: integer
      tags:
      - backoffice
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedPageBackofficeList'
          description: ''
    post:
      operationId: backoffice_pages_create
      description: CRUD + workflow éditorial des pages — `/api/v1/backoffice/pages`.
      tags:
      - backoffice
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PageBackoffice'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PageBackoffice'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PageBackoffice'
        required: true
      security:
      - jwtAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageBackoffice'
          description: ''
...
  /api/v1/backoffice/statistiques/concours:
    get:
      operationId: backoffice_statistiques_concours_retrieve
      description: |-
        GET /api/v1/backoffice/statistiques/concours?concours= — tableau de bord
        thématique Concours (EF-1102), calculé à la volée sur `apps.concours`.
      parameters:
      - in: query
        name: concours
        schema:
          type: string
        description: Code du concours (filtre).
      tags:
      - backoffice
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatistiquesConcours'
          description: ''
...
  /api/v1/concours:
    get:
      operationId: concours_list
      description: GET /api/v1/concours?q= — avis de concours ouverts (§7.4).
      parameters:
      - name: cursor
        required: false
        in: query
        description: La valeur du curseur de pagination.
        schema:
          type: string
      - name: limit
        required: false
        in: query
        description: Nombre de résultats à retourner par page.
        schema:
          type: integer
      tags:
      - concours
      security:
      - jwtAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedConcoursList'
          description: ''
...
components:
  schemas:
...
    ArticleBackoffice:
      type: object
      description: |-
        CRUD éditorial complet — auteur/dates en lecture seule, reste modifiable.

        `image_url` est calculée (jamais stockée telle quelle) : voir `Article.image` et
        `ArticleBackofficeViewSet.televerser_image`.
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        titre:
          type: string
          maxLength: 200
        slug:
          type: string
          maxLength: 220
          pattern: ^[-a-zA-Z0-9_]+$
        chapo:
          type: string
          description: Résumé affiché dans les listes.
        contenu:
          type: string
        statut:
          allOf:
          - $ref: '#/components/schemas/StatutContenuEnum'
          readOnly: true
        rubrique:
          type:
          - string
          - 'null'
          format: uuid
        date_publication:
          type:
          - string
          - 'null'
          format: date-time
          readOnly: true
        image_url:
          type: string
          readOnly: true
        galerie:
          type:
          - string
          - 'null'
          format: uuid
          description: Galerie d'images/vidéos complémentaires affichée sous l'article.
        meta_titre:
          type: string
          maxLength: 200
        meta_description:
          type: string
          maxLength: 300
        cree_le:
          type: string
          format: date-time
          readOnly: true
        modifie_le:
          type: string
          format: date-time
          readOnly: true
      required:
      - cree_le
      - date_publication
      - id
      - image_url
      - modifie_le
      - slug
      - statut
      - titre
...
    Concours:
      type: object
      description: Lecture publique — avis de concours.
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        titre:
          type: string
          readOnly: true
        code:
          type: string
          readOnly: true
          pattern: ^[-a-zA-Z0-9_]+$
        description:
          type: string
          readOnly: true
        conditions:
          type: string
          readOnly: true
          description: Conditions d'accès, pièces exigées.
        frais_inscription:
          type: string
          format: decimal
          pattern: ^-?\d{0,10}(?:\.\d{0,0})?$
          readOnly: true
          description: Montant en FCFA.
        date_ouverture:
          type: string
          format: date
          readOnly: true
        date_cloture:
          type: string
          format: date
          readOnly: true
        date_concours:
          type:
          - string
          - 'null'
          format: date
          readOnly: true
        places_disponibles:
          type:
          - integer
          - 'null'
          readOnly: true
        statut:
          allOf:
          - $ref: '#/components/schemas/StatutConcoursEnum'
          readOnly: true
        document_avis_url:
          type: string
          readOnly: true
      required:
      - code
      - conditions
      - date_cloture
      - date_concours
      - date_ouverture
      - description
      - document_avis_url
      - frais_inscription
      - id
      - places_disponibles
      - statut
      - titre
...
    PageBackoffice:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        titre:
          type: string
          maxLength: 200
        slug:
          type: string
          maxLength: 220
          pattern: ^[-a-zA-Z0-9_]+$
        contenu:
          type: string
        statut:
          allOf:
          - $ref: '#/components/schemas/StatutContenuEnum'
          readOnly: true
        rubrique:
          type:
          - string
          - 'null'
          format: uuid
        meta_titre:
          type: string
          maxLength: 200
        meta_description:
          type: string
          maxLength: 300
        cree_le:
          type: string
          format: date-time
          readOnly: true
        modifie_le:
          type: string
          format: date-time
          readOnly: true
      required:
      - cree_le
      - id
      - modifie_le
      - slug
      - statut
      - titre
...
    PaginatedArticleBackofficeList:
      type: object
      required:
      - results
      properties:
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?cursor=cD00ODY%3D"
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?cursor=cj0xJnA9NDg3
        results:
          type: array
          items:
            $ref: '#/components/schemas/ArticleBackoffice'
...
    PaginatedConcoursList:
      type: object
      required:
      - results
      properties:
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?cursor=cD00ODY%3D"
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?cursor=cj0xJnA9NDg3
        results:
          type: array
          items:
            $ref: '#/components/schemas/Concours'
...
    PaginatedPageBackofficeList:
      type: object
      required:
      - results
      properties:
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?cursor=cD00ODY%3D"
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?cursor=cj0xJnA9NDg3
        results:
          type: array
          items:
            $ref: '#/components/schemas/PageBackoffice'
...
    Repartition:
      type: object
      properties:
        cle:
          type: string
        total:
          type: integer
      required:
      - cle
      - total
    RepartitionMensuelle:
      type: object
      properties:
        mois:
          type: string
          description: Format AAAA-MM.
        total:
          type: integer
      required:
      - mois
      - total
...
    StatistiquesConcours:
      type: object
      properties:
        total:
          type: integer
        par_statut:
          type: array
          items:
            $ref: '#/components/schemas/Repartition'
        par_concours:
          type: array
          items:
            $ref: '#/components/schemas/Repartition'
        par_mois:
          type: array
          items:
            $ref: '#/components/schemas/RepartitionMensuelle'
      required:
      - par_concours
      - par_mois
      - par_statut
      - total
...
    StatutConcoursEnum:
      enum:
      - BROUILLON
      - OUVERT
      - CLOTURE
      - RESULTATS_PUBLIES
      type: string
      description: |-
        * `BROUILLON` - Brouillon
        * `OUVERT` - Ouvert
        * `CLOTURE` - Clôturé
        * `RESULTATS_PUBLIES` - Résultats publiés
    StatutContenuEnum:
      enum:
      - BROUILLON
      - RELECTURE
      - VALIDE
      - PUBLIE
      - ARCHIVE
      type: string
      description: |-
        * `BROUILLON` - Brouillon
        * `RELECTURE` - En relecture
        * `VALIDE` - Validé
        * `PUBLIE` - Publié
        * `ARCHIVE` - Archivé
...
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT